feature: remove unused code and move converter to corresponding package.
