--- scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/EnvConverter.java
@@ -1,5 +1,0 @@
-package cn.sliew.scaleph.plugin.seatunnel.flink;
-
-public interface EnvConverter extends ConfConverter {
-
-}
