--- scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/conf/BaseSinkConnectorOptions.java
@@ -1,13 +1,0 @@
-package cn.sliew.scaleph.plugin.seatunnel.flink.conf;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class BaseSinkConnectorOptions {
-
-    @JsonProperty("source_table_name")
-    private String sourceTableName;
-}
