--- scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/conf/SeatunnelEnvironment.java
@@ -1,6 +1,0 @@
-package cn.sliew.scaleph.plugin.seatunnel.flink.conf;
-
-public class SeatunnelEnvironment {
-
-
-}
