--- scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/conf/package-info.java
@@ -1,4 +1,0 @@
-/**
- * todo composite pattern
- */
-package cn.sliew.scaleph.plugin.seatunnel.flink.conf;
