+++ scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/converter/EnvConverter.java
@@ -1,0 +1,5 @@
+package cn.sliew.scaleph.plugin.seatunnel.flink.converter;
+
+public interface EnvConverter extends ConfConverter {
+
+}
