+++ scaleph-plugins/scaleph-plugin-seatunnel-native-flink/src/main/java/cn/sliew/scaleph/plugin/seatunnel/flink/converter/SourceConverter.java
@@ -1,4 +1,4 @@
-package cn.sliew.scaleph.plugin.seatunnel.flink;
+package cn.sliew.scaleph.plugin.seatunnel.flink.converter;
 
 public interface SourceConverter extends ConfConverter {
